[python,ci-examples] Support SGX extended product ID
KSS is a new feature that introduces additional fields to identify
enclave in build time: ISVEXTPRODID and ISVFAMILYID. These fields are
defined the the developer and signed in SIGSTRUCT.
Add new fields to the manifest to add support, and support in SGX
signing tool.

Add a hello world example with KSS enabled:
Product ID fields are configured in the manifest and the enclave
prints them.

Signed-off-by: Lavi, Nir <nir.lavi@intel.com>